requirements.txt
cdk8s==2.7.56
constructs==10.2.70
pyyaml==6.0.2
